feat: implement blog detail page with full content display
- add detail page for individual blog posts
- display full blog content including title, author, and publication date
- ensure layout includes likes and comments count
- style the page to match the overview page's design
