--- src/app/app.component.scss
@@ -1,120 +1,0 @@
-.page-title {
-  text-align: center;
-  font-weight: bold;
-  margin-bottom: 32px;
-}
-
-.create-blog-button {
-  position: absolute;
-  top: 16px;
-  right: 16px;
-}
-
-.no-blogs {
-  text-align: center;
-  font-size: 1.2em;
-  color: #757575;
-}
-
-.blog-navigation {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 16px;
-}
-
-.blog-container {
-  display: flex;
-  gap: 16px;
-  overflow: hidden;
-  width: 100%;
-  max-width: 1000px;
-  justify-content: center;
-  flex-wrap: wrap;
-}
-
-.blog-card {
-  width: 300px;
-  flex-shrink: 0;
-  margin: 10px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
-  border-radius: 8px;
-}
-
-.blog-details {
-  margin-top: 40px;
-  display: flex;
-  justify-content: center;
-}
-
-.detailed-blog-card {
-  width: 90%;
-  max-width: 800px;
-  margin: 0 auto;
-  padding: 20px;
-  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.3);
-  border-radius: 8px;
-}
-
-.new-blog-form {
-  margin-top: 20px;
-  display: flex;
-  justify-content: center;
-}
-
-mat-card-title {
-  font-weight: bold;
-}
-
-/* Like-Button Animation */
-.like-button {
-  color: #1976d2;
-  transition:
-    color 0.3s ease,
-    transform 0.1s ease;
-}
-
-.like-button.liked {
-  color: #ff4081;
-  animation: likeAnimation 0.3s ease-in-out;
-}
-
-@keyframes likeAnimation {
-  0% {
-    transform: scale(1);
-  }
-  50% {
-    transform: scale(1.2);
-  }
-  100% {
-    transform: scale(1);
-  }
-}
-
-button[mat-icon-button]:hover {
-  color: #1565c0;
-}
-
-mat-card-actions {
-  display: flex;
-  justify-content: space-between;
-}
-
-/* Responsive Design für kleine Bildschirme */
-@media (max-width: 600px) {
-  .blog-card {
-    width: 100%;
-    margin: 10px 0;
-  }
-
-  .blog-navigation {
-    flex-direction: column;
-  }
-}
-
-/* Responsive Design für Tablets */
-@media (max-width: 768px) {
-  .blog-card {
-    width: 45%;
-  }
-}
