+++ src/app/features/blog-details-page/blog-details-page.component.html
@@ -1,0 +1,22 @@
+<div class="blog-details-container" *ngIf="blog">
+  <div class="blog-header">
+    <h1>{{ blog.title }}</h1>
+    <p class="author">Autor: {{ blog.author }}</p>
+    <p class="date">Veröffentlicht am: {{ blog.createdAt | date }}</p>
+  </div>
+
+  <div class="blog-content">
+    <img
+      *ngIf="blog.headerImageUrl"
+      [src]="blog.headerImageUrl"
+      alt="Header Image"
+      class="header-image"
+    />
+    <p [innerHTML]="blog?.contentPreview"></p>
+  </div>
+
+  <div class="blog-stats">
+    <p><strong>Likes:</strong> {{ blog.likes }}</p>
+    <p><strong>Kommentare:</strong> {{ blog.comments }}</p>
+  </div>
+</div>
