+++ src/app/features/blog-details-page/blog-details-page.component.scss
@@ -1,0 +1,49 @@
+.blog-details-container {
+  max-width: 800px;
+  margin: auto;
+  padding: 20px;
+  background-color: #ffffff;
+  border-radius: 8px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+.blog-header {
+  text-align: center;
+
+  h1 {
+    font-size: 2.5rem;
+    margin-bottom: 0.5rem;
+  }
+
+  .author,
+  .date {
+    color: #555;
+    font-size: 0.9rem;
+  }
+}
+
+.header-image {
+  width: 100%;
+  max-height: 400px;
+  object-fit: cover;
+  border-radius: 8px;
+  margin: 20px 0;
+}
+
+.blog-content {
+  font-size: 1.1rem;
+  line-height: 1.6;
+  color: #333;
+}
+
+.blog-stats {
+  display: flex;
+  justify-content: space-around;
+  margin-top: 20px;
+  font-size: 1rem;
+  color: #333;
+
+  p {
+    margin: 0;
+  }
+}
