+++ src/app/features/blog-details-page/blog-details-page.component.spec.ts
@@ -1,0 +1,22 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { BlogDetailsPageComponent } from './blog-details-page.component';
+
+describe('BlogDetailsPageComponent', () => {
+  let component: BlogDetailsPageComponent;
+  let fixture: ComponentFixture<BlogDetailsPageComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [BlogDetailsPageComponent],
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(BlogDetailsPageComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
