+++ src/app/features/blog-overview-page/blog-overview-page.component.html
@@ -1,0 +1,73 @@
+<h1 class="page-title">{{ title }}</h1>
+
+<!-- Button zum Erstellen eines neuen Blogs -->
+<div class="create-blog-button">
+  <button
+    class="new-blog-button"
+    mat-raised-button
+    color="primary"
+    (click)="startCreatingNewBlog()"
+    aria-label="Neuen Blog erstellen"
+  >
+    Neuer Blog
+  </button>
+</div>
+
+<!-- Nachricht anzeigen, wenn keine Blogs verfügbar sind -->
+<div *ngIf="blogs.length === 0 && !isLoading" class="no-blogs">
+  <p>Keine Blogs verfügbar.</p>
+</div>
+
+<!-- Ladeindikator -->
+<mat-spinner *ngIf="isLoading" diameter="40"></mat-spinner>
+
+<!-- Navigationspfeile -->
+<div class="blog-navigation" *ngIf="blogs.length > 0 && !isLoading">
+  <button
+    class="nav-button"
+    mat-icon-button
+    (click)="prevBlog()"
+    [disabled]="currentIndex === 0"
+    aria-label="Vorheriger Blog"
+  >
+    <mat-icon>chevron_left</mat-icon>
+  </button>
+
+  <!-- Blogkarten -->
+  <div class="blog-container">
+    <div class="blog-card" *ngFor="let blog of displayedBlogs">
+      <img
+        class="blog-image"
+        [src]="blog.headerImageUrl"
+        alt="Header Image"
+        *ngIf="blog.headerImageUrl"
+      />
+
+      <div class="blog-content">
+        <span class="blog-tag">{{ blog.author }}</span>
+        <h2 class="blog-title">{{ blog.title }}</h2>
+        <p class="blog-preview">{{ blog.contentPreview }}</p>
+        <div class="blog-info">
+          <p><strong>Likes:</strong> {{ blog.likes }}</p>
+          <p><strong>Kommentare:</strong> {{ blog.comments }}</p>
+        </div>
+        <button
+          class="read-more-button"
+          [routerLink]="['/blog-details', blog.id]"
+        >
+          Mehr lesen
+        </button>
+      </div>
+    </div>
+  </div>
+
+  <button
+    class="nav-button"
+    mat-icon-button
+    (click)="nextBlog()"
+    [disabled]="currentIndex + maxBlogs >= blogs.length"
+    aria-label="Nächster Blog"
+  >
+    <mat-icon>chevron_right</mat-icon>
+  </button>
+</div>
