+++ src/app/features/blog-overview-page/blog-overview-page.routes.ts
@@ -1,0 +1,10 @@
+import { BlogOverviewPageComponent } from './blog-overview-page.component';
+import { Routes } from '@angular/router';
+
+const BLOG_OVERVIEW_PAGE_ROUTES: Routes = [
+  {
+    path: '',
+    component: BlogOverviewPageComponent,
+  },
+];
+export default BLOG_OVERVIEW_PAGE_ROUTES;
